First attempt at moving reasoning code into App.vue.
